fix cfindrq bug, cfindscp can receive the command of cfind-PDU

+++ dimse/dimse.h
@@ -31,12 +31,12 @@ public:
     
 public:
     void InitElementTag(DcmElement *dcmelement, uint16_t group, uint16_t element);
-    void InitElementData(DcmElement *dcmelement, uint16_t len, const unsigned char * data);
+    void InitElementData(DcmElement *dcmelement, uint32_t len, const unsigned char * data);
     void InitElementData(DcmElement *dcmelement, uint32_t data);
     void InitElementData(DcmElement *dcmelement, uint16_t data);
     void InitElementData(DcmElement *dcmelement, uint8_t data);
 private:
-    void InitElementData(DcmElement *dcmelement, int len, void *data);
+    void InitElementData(DcmElement *dcmelement, uint32_t len, void *data);
 public:
     DcmElement  groupLength;
     DcmElement  affectedSOPClassUID;
